+++ semantic-conventions/src/tests/semconv/model/test_correct_parse.py
@@ -735,6 +735,24 @@ class TestCorrectParse(unittest.TestCase):
         self.assertEqual(expected["n_constraints"], len(s.constraints))
         self.assertEqual(expected["attributes"], [a.fqn for a in s.attributes])
 
+    def test_scope_attribute(self):
+        semconv = SemanticConventionSet(debug=False)
+        semconv.parse(self.load_file("yaml/scope.yaml"))
+        self.assertEqual(len(semconv.models), 1)
+
+        expected = {
+            "id": "scope-id",
+            "prefix": "",
+            "type": "scope",
+            "extends": "",
+            "brief": "Instrumentation Scope attributes",
+            "n_constraints": 0,
+            "attributes": [
+                "short_name",
+            ],
+        }
+        self.semantic_convention_check(list(semconv.models.values())[0], expected)
+
     def metric_check(self, metrics: Tuple[MetricSemanticConvention.Metric], expected):
         for m, ex in zip(metrics, expected):
             self.assertEqual(m.id, ex["id"])